Shuffled dataset, tfrecord validation, silence

+++ data/tfrecord_writer.py
@@ -28,7 +28,7 @@ class TFrecord_Writer:
                 print "\r", 'Converting wavs: {}/{}'.format(i, n_wavs),
 
             wav = self.dataset[i].get_wav()
-            wav = wav.astype(np.float32)
+            # wav = wav.astype(np.float32)
 
             label = self.dataset[i].label.index
 
@@ -37,9 +37,6 @@ class TFrecord_Writer:
                 'wav': _float_feature_array(wav)
             }
 
-            # example = tf.train.Example()
-            # example.features.feature["label"].int64_list.value.append(label)
-            # example.features.feature["wav"].int64_list.value.append(wav)
             example = tf.train.Example(features=tf.train.Features(feature=feature))
 
             writer.write(example.SerializeToString())
